[ELF] Don't emit weak undefined symbols in .dynsym for executables
Summary:
D59275 added the following clause to Symbol::includeInDynsym()

    if (isUndefWeak() && Config->Pie && SharedFiles.empty())
      return false;

It seems we don't have to differentiate non-PIE from PIE, thus we can
generalize it to:

    if (!Config->Shared && isUndefWeak() && SharedFiles.empty())
      return false;

After D59549, we essentially treat relocations to weak-undef as
link-time constants when producing an executable. So, simplify the above
further to:

    if (!Config->Shared && isUndefWeak())
      return false;

ld.bfd has complex and architecture-varying rules. We try to provide
reasonable semantics to exclude some weak undefined symbols but not be
too smart.

In the updated tests, weak-undef-rw.s is folded into
weak-undef-exported.s as they are checking the same thing except that
the former checks a writable section.

Reviewers: ruiu, pcc, sivachandra, espindola

Subscribers: emaste, arichardson, jdoerfert, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D59549