[ELF] -r: fix R_*_NONE to section symbols on Elf*_Rel targets
On Elf*_Rel targets, for a relocation to a section symbol, an R_ABS is
added to compute the implicit addend which will be applied in the
relocated location.

Addends of R_*_NONE should be ignored, so don't emit an R_ABS.

This fixes crashes on X86 and ARM (their relocateOne() do not handle
R_*_NONE).

Reviewers: grimar, peter.smith, rprichard, ruiu, espindola

Reviewed By: peter.smith

Subscribers: emaste, javed.absar, arichardson, kristof.beyls, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D62052