[ELF] -r: fix crash when processing a SHT_REL[A] that relocates a SHF_MERGE after D67504/r372734
Summary:
Fix PR43767

In -r mode, when processing a SHT_REL[A] that relocates a SHF_MERGE, sec->getRelocatedSection() is a
MergeInputSection and its parent is an OutputSection but is asserted to
be a SyntheticSection (MergeSyntheticSection).

The code path is not exercised in non -r mode because the relocated
section is changed from MergeInputSection to InputSection.

Fix the crash by reordering the code to convert MergeInputSection before
returning.

Reviewers: grimar, peter.smith, ruiu, espindola

Reviewed By: peter.smith

Subscribers: emaste, arichardson, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D69364